backend/pyproject.toml
[tool.poetry]
name = "thesis-tracker-backend"
version = "0.1.0"
description = "Backend for Thesis Tracker application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
sqlalchemy = "^2.0.25"
alembic = "^1.12.1"
pydantic = {extras = ["email"], version = "^2.5.2"}
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.1.0"
email-validator = "^2.1.0"
httpx = "^0.27.0"
authlib = "^1.3.0"
bcrypt = "3.2.0"
python-docx = "^0.8.11"
PyPDF2 = "^3.0.1"
mammoth = "^1.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black" 